[mod] move hook to set Unix thread name into searx.unixthreadname
requires setproctitle (but optional)
