+++ searx/webapp.py
@@ -111,21 +111,6 @@ from searx.network import stream as http_stream
 from searx.search.checker import get_result as checker_get_result
 from searx.settings_loader import get_default_settings_path
 
-# set Unix thread name
-try:
-    import setproctitle
-except ImportError:
-    pass
-else:
-    import threading
-    old_thread_init = threading.Thread.__init__
-
-    def new_thread_init(self, *args, **kwargs):
-        # pylint: disable=protected-access, disable=c-extension-no-member
-        old_thread_init(self, *args, **kwargs)
-        setproctitle.setthreadtitle(self._name)
-    threading.Thread.__init__ = new_thread_init
-
 logger = logger.getChild('webapp')
 
 # check secret_key
